backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4"
dependencies = [
    "django (>=4.2,<5.0)",
    "djangorestframework (>=3.15.2,<4.0.0)",
    "pillow (>=11.1.0,<12.0.0)",
    "djangorestframework-simplejwt (>=5.4.0,<6.0.0)",
    "django-cors-headers (>=4.6.0,<5.0.0)",
    "django-filter (>=24.3,<25.0)",
    "drf-spectacular (>=0.28.0,<0.29.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "django-environ (>=0.12.0,<0.13.0)",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
ruff = "^0.9.2"
django-debug-toolbar = "^5.0.1"
ipython = "^8.31.0"
isort = "^5.13.2"
faker = "^36.1.1"
django-silk = "^5.3.2"

[tool.poetry]
package-mode = false

[tool.isort]
skip = [
"migrations",
"**/migrations",
]